requirements.txt
Cartopy==0.20.1
geopy==2.2.0
matplotlib==3.5.1
numpy==1.22.0
pandas==1.5.2
Pillow==9.5.0
PyKrige==1.6.1
python_dateutil==2.8.2
pytz==2021.3
Requests==2.30.0
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.1
suntime==1.2.5
tzlocal==4.1
